Revert "Arm template script iot resource"
