+++ README.md
@@ -1,0 +1,20 @@
+# ad440-winter2020-tuesday-repo
+
+Welcome to the North Seattle AD440 "Coffee" Cloud Computing Practicum! 
+
+<img src="https://food.fnr.sndimg.com/content/dam/images/food/fullset/2014/11/6/0/fnd_cat-latte-art.jpg.rend.hgtvcom.406.305.suffix/1415304438292.jpeg" title="Coffee Project">
+
+## Getting Started
+This practicum is geared Azure, and uses React, Javascript, and Python. This app does not use a traditional stack, like LAMP, or SEAN. Rather it breaks its tasks down into descrete files of code that execute using Azure services like Functions 
+
+### Installing
+While the app is deployed to the cloud, there are several local tools one should have local on their computer if they want to join in the development of this app. You may not need all the tools listed to work on this app.
+
+### Node.js
+https://nodejs.org/en/download/
+
+### React
+https://reactjs.org/
+
+### Python
+https://www.python.org/
