--- devops/azure-IotHub/deploy.ps1
@@ -1,10 +1,0 @@
-#lists all current IoT Hubs resources
-Get-AzIotHub
-
-
-$resourceGroupName = Read-Host -Prompt "Enter the Resource Group name"
-$location = Read-Host -Prompt "Enter the location (i.e. centralus)"
-$templateUri = "https://raw.githubusercontent.com/Yozelinm/Yozelinm.github.io/master/iottemplate.json"
-New-AzResourceGroupDeployment -ResourceGroupName $resourceGroupName -TemplateUri $templateUri -Location $location
-
-
