--- devops/azure-IotHub/parameters.json
@@ -1,10 +1,0 @@
-{
-    "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentParameters.json#",
-    "contentVersion": "1.0.0.0",
-    "parameters": {
-        "iotHubName": {
-            "value": "devstore"
-        }
-        
-    }
-}
