--- devops/azure-IotHub/template.json
@@ -1,83 +1,0 @@
-{
-  "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentTemplate.json#",
-  "contentVersion": "1.0.0.0",
-  "parameters": {
-    "iotHubName": {
-      "type": "string",
-      "minLength": 3,
-      "metadata": {
-        "description": "Specifies the name of the IoT Hub."
-      }
-    },
-    "location": {
-      "type": "string",
-      "defaultValue": "[resourceGroup().location]",
-      "metadata": {
-              "description": "Location for all resources."
-      }
-    },
-    "skuName": {
-      "type": "string",
-      "defaultValue": "F1",
-      "metadata": {
-        "description": "Specifies the IotHub SKU."
-      }
-    },
-    "capacityUnits": {
-      "type": "int",
-      "minValue": 1,
-      "maxValue": 1,
-      "defaultValue": 1,
-      "metadata": {
-        "description": "Specifies the number of provisioned IoT Hub units. Restricted to 1 unit for the F1 SKU. Can be set up to maximum number allowed for subscription."
-      }
-    }
-  },
-  "variables": {
-    "consumerGroupName": "[concat(parameters('iotHubName'), '/events/cg1')]"
-  },
-  "resources": [
-    {
-      "type": "Microsoft.Devices/IotHubs",
-      "apiVersion": "2018-04-01",
-      "name": "[parameters('iotHubName')]",
-      "location": "[parameters('location')]",
-      "properties": {
-        "eventHubEndpoints": {
-          "events": {
-            "retentionTimeInDays": 1,
-            "partitionCount": 2
-          }
-        },
-        "cloudToDevice": {
-          "defaultTtlAsIso8601": "PT1H",
-          "maxDeliveryCount": 10,
-          "feedback": {
-            "ttlAsIso8601": "PT1H",
-            "lockDurationAsIso8601": "PT60S",
-            "maxDeliveryCount": 10
-          }
-        },
-        "messagingEndpoints": {
-          "fileNotifications": {
-            "ttlAsIso8601": "PT1H",
-            "lockDurationAsIso8601": "PT1M",
-            "maxDeliveryCount": 10
-          }
-        }
-      },
-      "sku": {
-        "name": "[parameters('skuName')]",
-        "capacity": "[parameters('capacityUnits')]"
-      }
-    },
-    {
-      "type": "Microsoft.Devices/iotHubs/eventhubEndpoints/ConsumerGroups",
-      "apiVersion": "2018-04-01",
-      "name": "[variables('consumerGroupName')]",
-      "dependsOn": [
-        "[resourceId('Microsoft.Devices/IotHubs', parameters('iotHubName'))]"
-      ]
-    }
-  ]
-}
