+++ devops/azure-StreamAnalytics/.gitkeep
@@ -1,0 +1,1 @@
+
