+++ ui/src/components/AdalConfig.js
@@ -1,0 +1,4 @@
+// azure active directory client ID
+export default {
+  clientId: 'ecb37f55-02d0-4cab-902d-b7aaaecca80e',
+}
