+++ ui/src/css/App.css
@@ -24,7 +24,7 @@
 
 .App-header {
   background-color: #282c34;
-  min-height: 100vh;
+  min-height: 10vh;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -45,8 +45,6 @@
     transform: rotate(360deg);
   }
 }
-<<<<<<< HEAD:ui/react/src/App.css
-=======
 .login {
   background-color: #282c34;
 }
@@ -73,4 +71,3 @@
 
 
 
->>>>>>> develop:ui/src/css/App.css
