+++ ui/src/css/GlobalFilter.css
