+++ ui/src/css/MachinesTable.css
