+++ ui/src/index.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './css/index.css';
 import App from './App';
 
 import * as serviceWorker from './serviceWorker';
-import ToggleSwitch from './components/toggleSwitch';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-ReactDOM.render(<ToggleSwitch />, document.getElementById('root'));
-
+//ReactDOM.render(<ToggleSwitch />, document.getElementById('root'));
+//ReactDOM.render(< TextComponent/>, document.getElementById('root'));
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
 // Learn more about service workers: https://bit.ly/CRA-PWA
